edit models and migrations files

+++ database/migrations/2022_03_17_230640_create_restaurants_table.php
@@ -20,7 +20,7 @@ class CreateRestaurantsTable extends Migration {
 			$table->string('whatsapp')->unique();
 			$table->string('image');
 			$table->enum('status', array('open', 'close'));
-			$table->integer('is_active')->default('1');
+			$table->tinyInteger('is_active')->default('1');
 			$table->string('api_token', 60)->unique()->nullable();
 			$table->string('pin_code')->nullable();
 			$table->timestamps();
